+++ lib/src/profile/domain/profile_params.dart
@@ -3,6 +3,6 @@ import 'package:tracksu/src/profile/domain/profile_user_reference.dart';
 
 final class ProfileParams {
   const ProfileParams({required this.user, required this.ruleset});
-  final ProfileUserId user;
+  final ProfileUserReference user;
   final ProfileRuleset ruleset;
 }